Move layout to CombatScene.
