+++ src/controllers/combat_scene_controller_test.py
@@ -17,7 +17,7 @@ def create_combat_controller(enemy):
 
 
 def select_enemy(enemy, controller: CombatSceneController):
-    rects = controller._view.layout.get_rects(enemy)
+    rects = controller.scene.layout.get_rects(enemy)
     assert len(rects) == 1
 
     simulate_mouse_click(*rects[0].center)
