+++ src/views/artists/combat_options_artist.py
@@ -6,14 +6,12 @@ from models.scenes.combat_scene import CombatScene
 from models.scenes.scenes_base import Scene
 from views.artists.drawing_utils import rescale_horizontal, rescale_vertical
 from views.artists.scene_artist_base import SceneArtist
-from views.layouts import Layout
 from views.pygame_screen import Screen
 
 
 class CombatOptionsArtist(SceneArtist):
 
-    def render(self, screen: Screen, scene: Scene,
-               layout: Layout) -> None:
+    def render(self, screen: Screen, scene: Scene) -> None:
         assert isinstance(scene, CombatScene)
         options = self._combat_options(scene.current_moves)
         x, font_size, spacing = rescale_horizontal(450, 35, 50)
