+++ src/views/artists/inventory_artist.py
@@ -3,14 +3,12 @@ from data.constants import SCREEN_SIZE
 from models.scenes.inventory_scene import InventoryScene
 from models.scenes.scenes_base import Scene
 from views.artists.scene_artist_base import SceneArtist
-from views.layouts import Layout
 from views.pygame_screen import Screen
 
 
 class InventoryArtist(SceneArtist):
 
-    def render(self, screen: Screen, scene: Scene,
-               layout: Layout) -> None:
+    def render(self, screen: Screen, scene: Scene) -> None:
         assert isinstance(scene, InventoryScene)
 
         screen.render_texts(list(scene.options),
