+++ src/views/artists/scene_artist_base.py
@@ -8,5 +8,5 @@ from views.pygame_screen import Screen
 class SceneArtist(metaclass=abc.ABCMeta):
 
     @abc.abstractmethod
-    def render(self, screen: Screen, scene: Scene, layout: Layout) -> None:
+    def render(self, screen: Screen, scene: Scene) -> None:
         """Generic function to render content to a screen."""
