+++ src/views/artists/settings_artist.py
@@ -4,14 +4,12 @@ from models.scenes.scenes_base import Scene
 from models.scenes.settings_scene import SettingsScene
 from views.artists.drawing_utils import rescale_horizontal, rescale_vertical
 from views.artists.scene_artist_base import SceneArtist
-from views.layouts import Layout
 from views.pygame_screen import Screen
 
 
 class SettingsArtist(SceneArtist):
 
-    def render(self, screen: Screen, scene: Scene,
-               layout: Layout) -> None:
+    def render(self, screen: Screen, scene: Scene) -> None:
         assert isinstance(scene, SettingsScene)
 
         x, font_size, spacing = rescale_horizontal(250, 35, 50)
