+++ src/views/scene_view.py
@@ -1,8 +1,6 @@
-from functools import reduce
 from typing import List
 
 from data.colors import GREEN
-from data.constants import SCREEN_SIZE
 from models.scenes.combat_scene import CombatScene
 from models.scenes.decision_scene import DecisionScene
 from models.scenes.inventory_scene import InventoryScene
@@ -17,7 +15,6 @@ from views.artists.inventory_artist import InventoryArtist
 from views.artists.overlay_artist import OverlayArtist
 from views.artists.scene_artist_base import SceneArtist
 from views.artists.settings_artist import SettingsArtist
-from views.layouts import Layout
 from views.pygame_screen import get_screen
 
 
@@ -27,22 +24,18 @@ class SceneView(object):
         self._scene = scene
         self._screen = get_screen()
         self._artists = _build_scene_artists(scene)
-        self._layout = _build_scene_layout(scene)
         self._debug_mode = False
 
     def update(self) -> None:
         self._screen.clear()
         for artist in self._artists:
-            artist.render(self._screen, self._scene, self._layout)
+            artist.render(self._screen, self._scene)
         # VERY IMPORTANT TO CALL UPDATE ONCE
         if self._debug_mode:
-            for rect in self._layout.get_rects(self._layout):
+            assert isinstance(self._scene, CombatScene)
+            for rect in self._scene.layout.get_rects(self._scene.layout):
                 self._screen.render_rect(rect, GREEN, 2)
         self._screen.update()
-
-    @property
-    def layout(self) -> Layout:
-        return self._layout
 
     def toggle_debug(self) -> None:
         """Toggle DEBUG mode, where all layout rects are drawn."""
@@ -63,30 +56,3 @@ def _build_scene_artists(scene: Scene) -> List[SceneArtist]:
     else:
         raise ValueError('Unrecognized Scene {}'.format(scene))
     return artists
-
-
-def _build_scene_layout(scene: Scene) -> Layout:
-    if isinstance(scene, CombatScene):
-        characters = scene.characters()
-        # player side layout
-        player = characters[0]
-
-        player_layout = Layout([(None, 2), (player, 1), (None, 2)], 'vertical')
-        player_layout = Layout([(None, 1), (player_layout, 1), (None, 1)],
-                               'horizontal')
-
-        # stack layout
-        stack_layout = Layout()
-
-        # enemies layout
-        assert len(characters) > 1
-        elements = reduce(lambda a, b: a + b,
-                          ([(None, 1), (e, 1)] for e in characters[1:]))
-        elements.append((None, 1))
-        enemies_layout = Layout(elements, 'vertical')
-
-        return Layout(
-            [(player_layout, 1), (stack_layout, 1), (enemies_layout, 1)],
-            'horizontal', SCREEN_SIZE)
-
-    return Layout(dimensions=SCREEN_SIZE)
